thread-local arena: grow buf if there is no enough space (#116)
* Arena: add UT for arena
* Arena: grow buffer if there is no enough space, so Arena::alloc alway will success
* Arena: Allocate 8-byte aligned memory
* Arena: support larger arena whose size is large than 4GB

+++ src/db.rs
@@ -62,7 +62,7 @@ impl Core {
         let path = Self::memtable_file_path(opts, file_id);
         let c = make_comparator();
         // TODO: refactor skiplist to use `u64`
-        let skl = Skiplist::with_capacity(c, opts.arena_size() as u32);
+        let skl = Skiplist::with_capacity(c, opts.arena_size() as usize);
 
         // We don't need to create the WAL for the skiplist in in-memory mode so return the memtable.
         if opts.in_memory {
